+++ database/cockroachdb/cockroachdb.go
@@ -225,7 +225,7 @@ func (c *CockroachDb) Run(migration io.Reader) error {
 	return nil
 }
 
-func (c *CockroachDb) SetVersion(version int, dirty bool) error {
+func (c *CockroachDb) SetVersion(version int64, dirty bool) error {
 	return crdb.ExecuteTx(context.Background(), c.db, nil, func(tx *sql.Tx) error {
 		if _, err := tx.Exec(`DELETE FROM "` + c.config.MigrationsTable + `"`); err != nil {
 			return err
@@ -244,7 +244,7 @@ func (c *CockroachDb) SetVersion(version int, dirty bool) error {
 	})
 }
 
-func (c *CockroachDb) Version() (version int, dirty bool, err error) {
+func (c *CockroachDb) Version() (version int64, dirty bool, err error) {
 	query := `SELECT version, dirty FROM "` + c.config.MigrationsTable + `" LIMIT 1`
 	err = c.db.QueryRow(query).Scan(&version, &dirty)
