+++ database/redshift/redshift.go
@@ -209,7 +209,7 @@ func runesLastIndex(input []rune, target rune) int {
 	return -1
 }
 
-func (p *Redshift) SetVersion(version int, dirty bool) error {
+func (p *Redshift) SetVersion(version int64, dirty bool) error {
 	tx, err := p.conn.BeginTx(context.Background(), &sql.TxOptions{})
 	if err != nil {
 		return &database.Error{OrigErr: err, Err: "transaction start failed"}
@@ -243,7 +243,7 @@ func (p *Redshift) SetVersion(version int, dirty bool) error {
 	return nil
 }
 
-func (p *Redshift) Version() (version int, dirty bool, err error) {
+func (p *Redshift) Version() (version int64, dirty bool, err error) {
 	query := `SELECT version, dirty FROM "` + p.config.MigrationsTable + `" LIMIT 1`
 	err = p.conn.QueryRowContext(context.Background(), query).Scan(&version, &dirty)
 	switch {
